Поменял конфиг на cleanenv
