+++ README.md
@@ -6,6 +6,8 @@ Authorization: jwt
 
 Validation: https://github.com/go-ozzo/ozzo-validation
 
+Config: https://github.com/ilyakaznacheev/cleanenv
+
 HTTP framework: https://github.com/gorilla/mux
 
 DB: mongodb (go.mongodb.org/mongo-driver/mongo)
