--- config/app.yaml
@@ -1,6 +1,0 @@
-host: "localhost"
-port: "27017"
-user: "root"
-pass: "example"
-signingkey: "example"
-salt: "example"
