+++ internal/domain/entity/user.go
@@ -1,7 +1,7 @@
 package entity
 
 import (
-	"ChatGo/config"
+	"ChatGo/internal/config"
 	"crypto/md5"
 	validation "github.com/go-ozzo/ozzo-validation/v4"
 	"github.com/go-ozzo/ozzo-validation/v4/is"
